Split up the Players module
